+++ octasine/src/audio/parameters/operator_frequency_fine.rs
@@ -18,13 +18,13 @@ impl AudioParameter for OperatorFrequencyFineAudioParameter {
     }
     fn get_value_with_lfo_addition(
         &mut self,
-        lfo_addition: Option<f64>,
+        lfo_addition: Option<f32>,
     ) -> <Self::ParameterValue as ParameterValue>::Value {
         if let Some(lfo_addition) = lfo_addition {
             // log2(1.5) / 2
-            const FACTOR: f64 = 0.5849625007211562 / 2.0;
+            const FACTOR: f32 = 0.5849625007211562 / 2.0;
 
-            self.get_value() * 2.0f64.powf(FACTOR * lfo_addition)
+            self.get_value() * 2.0f32.powf(FACTOR * lfo_addition) as f64
         } else {
             self.get_value()
         }
