+++ octasine/src/parameters/operator_wave_type.rs
@@ -14,7 +14,7 @@ impl Default for WaveType {
     }
 }
 impl CalculateCurve for WaveType {
-    fn calculate(self, phase: Phase) -> f64 {
+    fn calculate(self, phase: Phase) -> f32 {
         match self {
             Self::Sine => crate::parameters::lfo_shape::sine(phase),
             Self::WhiteNoise => {
@@ -23,7 +23,7 @@ impl CalculateCurve for WaveType {
                 // its algorithm.
                 let seed = phase.0.to_bits() + 2;
 
-                (fastrand::Rng::with_seed(seed).f64() - 0.5) * 2.0
+                (fastrand::Rng::with_seed(seed).f32() - 0.5) * 2.0
             }
         }
     }
